test: Splits balance tests and removes force clicking
