+++ frontend/app/tests/e2e/pages/rotki-app.ts
@@ -70,13 +70,12 @@ export class RotkiApp {
       '[data-cy="privacy-mode-dropdown__input"] ~ .v-slider__thumb-container'
     ).as('input');
 
-    cy.get('@input').focus().type('{downarrow}'.repeat(2), { force: true });
+    cy.get('@input').focus().type('{downarrow}'.repeat(2));
 
     if (mode > 0) {
-      cy.get('@input').type('{uparrow}'.repeat(mode), { force: true });
+      cy.get('@input').type('{uparrow}'.repeat(mode));
     }
     cy.get('.privacy-mode-dropdown').click();
-    cy.get('.user-dropdown').click();
   }
 
   drawerIsVisible(isVisible: boolean) {
