+++ frontend/app/tests/e2e/specs/settings.spec.ts
@@ -125,6 +125,13 @@ describe('settings', () => {
     });
 
     describe('ignored asset settings', () => {
+      let ignoredAssets = 0;
+      before(() => {
+        pageAccounting.ignoredAssets().then(ignored => {
+          ignoredAssets = parseInt(ignored);
+        });
+      });
+
       it('add an ignored asset and validate UI message it has been added', () => {
         pageAccounting.addIgnoredAsset('1SG');
         pageAccounting.confirmInlineSuccess(
@@ -132,7 +139,7 @@ describe('settings', () => {
           '1SG'
         );
       });
-      it('add another 2 ignored assets and confirm count is 4', () => {
+      it('add another 2 ignored assets and confirm count increased by 3', () => {
         pageAccounting.addIgnoredAsset('ZIX');
         pageAccounting.confirmInlineSuccess(
           '.accounting-settings__asset-to-ignore',
@@ -143,15 +150,17 @@ describe('settings', () => {
           '.accounting-settings__asset-to-ignore',
           '1CR'
         );
-        pageAccounting.ignoredAssetCount('19');
+        // @ts-ignore
+        pageAccounting.ignoredAssetCount(ignoredAssets + 3);
       });
-      it('remove an ignored asset, validate UI message, and confirm count is 3', () => {
+      it('remove an ignored asset, validate UI message, and confirm count decreased by one', () => {
         pageAccounting.remIgnoredAsset('1SG');
         pageAccounting.confirmInlineSuccess(
           '.accounting-settings__ignored-assets',
           '1SG'
         );
-        pageAccounting.ignoredAssetCount('18');
+        // @ts-ignore
+        pageAccounting.ignoredAssetCount(ignoredAssets + 2);
       });
     });
   });
